Added `envvars.json` check

+++ builder/platform-uploads/opt/elasticbeanstalk/hooks/restartappserver/enact/03-apply-config.sh
@@ -11,29 +11,31 @@
 
 . /etc/TecdonorPlatform/platform.config
 
-## NGINX OPTIONS
-# DOCUMENT_ROOT=/var/www/html/`cat $CONFIG_DIR/envvars.json | jq -r '."aws:elasticbeanstalk:container:php:phpini".document_root | select (.!=null)'`
-# DEFAULT=`cat /etc/nginx/sites-available/default | sed -E 's|root(\s*)(.*);|root\1'"$DOCUMENT_ROOT"';|'`
-# echo "$DEFAULT" > /etc/nginx/sites-available/default
+if [ -f $CONFIG_DIR/envvars.json ]; then
+  ## NGINX OPTIONS
+  # DOCUMENT_ROOT=/var/www/html/`cat $CONFIG_DIR/envvars.json | jq -r '."aws:elasticbeanstalk:container:php:phpini".document_root | select (.!=null)'`
+  # DEFAULT=`cat /etc/nginx/sites-available/default | sed -E 's|root(\s*)(.*);|root\1'"$DOCUMENT_ROOT"';|'`
+  # echo "$DEFAULT" > /etc/nginx/sites-available/default
 
-## PHP OPTIONS
-MEMORY_LIMIT=`cat $CONFIG_DIR/envvars.json | jq -r .\"aws:elasticbeanstalk:container:php:phpini\".memory_limit`
-MAX_EXECUTION_TIME=`cat $CONFIG_DIR/envvars.json | jq -r .\"aws:elasticbeanstalk:container:php:phpini\".max_execution_time`
-DISPLAY_ERRORS=`cat $CONFIG_DIR/envvars.json | jq -r .\"aws:elasticbeanstalk:container:php:phpini\".display_errors`
-ALLOW_URL_FOPEN=`cat $CONFIG_DIR/envvars.json | jq -r .\"aws:elasticbeanstalk:container:php:phpini\".allow_url_fopen`
-ZLIB_OUTPUT_COMPRESSION=`cat $CONFIG_DIR/envvars.json | jq -r .\"aws:elasticbeanstalk:container:php:phpini\".zlib_output_compression`
+  ## PHP OPTIONS
+  MEMORY_LIMIT=`cat $CONFIG_DIR/envvars.json | jq -r .\"aws:elasticbeanstalk:container:php:phpini\".memory_limit`
+  MAX_EXECUTION_TIME=`cat $CONFIG_DIR/envvars.json | jq -r .\"aws:elasticbeanstalk:container:php:phpini\".max_execution_time`
+  DISPLAY_ERRORS=`cat $CONFIG_DIR/envvars.json | jq -r .\"aws:elasticbeanstalk:container:php:phpini\".display_errors`
+  ALLOW_URL_FOPEN=`cat $CONFIG_DIR/envvars.json | jq -r .\"aws:elasticbeanstalk:container:php:phpini\".allow_url_fopen`
+  ZLIB_OUTPUT_COMPRESSION=`cat $CONFIG_DIR/envvars.json | jq -r .\"aws:elasticbeanstalk:container:php:phpini\".zlib_output_compression`
 
-### PHP ENV
-cat $CONFIG_DIR/envvars.json | jq -r '."aws:elasticbeanstalk:application:environment"|to_entries|map("env[\(.key)]=\"\(.value|tostring)\"")|.[]|select(.!="env[]=\"null\"")' > /etc/php/7.0/fpm/env.conf
-REDIS_HOST=`cat $CONFIG_DIR/envvars.json | jq -r '."aws:elasticbeanstalk:application:environment".REDIS_HOST'`
-APP_URL=`cat $CONFIG_DIR/envvars.json | jq -r '."aws:elasticbeanstalk:application:environment".APP_URL'`
+  ### PHP ENV
+  cat $CONFIG_DIR/envvars.json | jq -r '."aws:elasticbeanstalk:application:environment"|to_entries|map("env[\(.key)]=\"\(.value|tostring)\"")|.[]|select(.!="env[]=\"null\"")' > /etc/php/7.0/fpm/env.conf
+  REDIS_HOST=`cat $CONFIG_DIR/envvars.json | jq -r '."aws:elasticbeanstalk:application:environment".REDIS_HOST'`
+  APP_URL=`cat $CONFIG_DIR/envvars.json | jq -r '."aws:elasticbeanstalk:application:environment".APP_URL'`
 
-if [ ! -z "$REDIS_HOST" ] && [ ! -z "$APP_URL" ]; then
-    cat /etc/laravel-echo-server-example.json | \
-        sed "s|__REDIS_HOST__|$REDIS_HOST|g" | \
-        sed "s|__APP_URL__|$APP_URL|g" > /etc/laravel-echo-server.json
+  if [ ! -z "$REDIS_HOST" ] && [ ! -z "$APP_URL" ]; then
+      cat /etc/laravel-echo-server-example.json | \
+          sed "s|__REDIS_HOST__|$REDIS_HOST|g" | \
+          sed "s|__APP_URL__|$APP_URL|g" > /etc/laravel-echo-server.json
 
-    supervisorctl update
-else
-    echo "REDIS_HOST and APP_URL environment variables must be set in order to have Laravel Echo Server working"
+      supervisorctl update
+  else
+      echo "REDIS_HOST and APP_URL environment variables must be set in order to have Laravel Echo Server working"
+  fi
 fi
